+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.8.2 (2023-11-04)
+* Deprecated execute() in favour of run()
+* Added default execute() to Bootstrap
+
 ## v0.8.1 (2023-11-04)
 * Return Kernel from initialize()